requirement.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
mysqlclient==1.3.12
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
urllib3==1.22
